blueprint decomposition
